requirements.txt
python-dotenv>=0.19.0,<0.20.0
uvloop>=0.15.0,<0.16.0
gino>=1.0.0,<1.1.0

git+https://github.com/Rapptz/discord.py
git+https://github.com/Lapis256/discord-ext-ui.git
